+++ presto-main/src/test/java/com/facebook/presto/operator/TestHashPartitionMaskOperator.java
@@ -94,10 +94,10 @@ public class TestHashPartitionMaskOperator
 
             MaterializedResult.Builder expected = resultBuilder(TEST_SESSION, BIGINT, BOOLEAN);
             for (int i = 0; i < ROW_COUNT; i++) {
-                int rawHash = (int) BigintOperators.hashCode(i);
+                long rawHash = BigintOperators.hashCode(i);
                 // mix the bits so we don't use the same hash used to distribute between stages
-                rawHash = (int) XxHash64.hash(Integer.reverse(rawHash));
-                rawHash &= Integer.MAX_VALUE;
+                rawHash = XxHash64.hash(Long.reverse(rawHash));
+                rawHash &= Long.MAX_VALUE;
 
                 boolean active = (rawHash % PARTITION_COUNT == partition);
                 expected.row((long) i, active);
@@ -138,10 +138,10 @@ public class TestHashPartitionMaskOperator
 
             MaterializedResult.Builder expected = resultBuilder(TEST_SESSION, BIGINT, BOOLEAN, BOOLEAN, BOOLEAN);
             for (int i = 0; i < ROW_COUNT; i++) {
-                int rawHash = (int) BigintOperators.hashCode(i);
+                long rawHash = BigintOperators.hashCode(i);
                 // mix the bits so we don't use the same hash used to distribute between stages
-                rawHash = (int) XxHash64.hash(Integer.reverse(rawHash));
-                rawHash &= Integer.MAX_VALUE;
+                rawHash = XxHash64.hash(Long.reverse(rawHash));
+                rawHash &= Long.MAX_VALUE;
 
                 boolean active = (rawHash % PARTITION_COUNT == partition);
                 boolean maskValue = i % 2 == 0;
