+++ presto-spi/src/main/java/com/facebook/presto/spi/block/AbstractArrayBlock.java
@@ -212,7 +212,7 @@ public abstract class AbstractArrayBlock
     }
 
     @Override
-    public int hash(int position, int offset, int length)
+    public long hash(int position, int offset, int length)
     {
         throw new UnsupportedOperationException();
     }
